add habit details view with button to delete
